pyproject.toml
[project]
name = "sharyo-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]
